# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: PrimeStream API
  version: 1.0.0
  description: API REST JWT pour tester PrimeStream et SUCCHERO avec Postman.
servers:
  - url: http://localhost:8091
    description: Serveur local
  - url: http://ADRESSE-IP-DU-SERVEUR:8091
    description: Acces depuis un autre PC du meme reseau
...
paths:
...
  /api/rest/profile/preferences:
    post:
      tags: [Profile]
      summary: Changer le theme intelligent et les preferences
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreferenceRequest'
            example:
              moodMode: romance
              preferredGenre: Romance
              preferredLanguage: Francais
              preferredQuality: Auto
              maturityLevel: 13+
      responses:
        '200':
          description: Profil mis a jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/rest/subscription:
    post:
      tags: [Profile]
      summary: Choisir un abonnement
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionRequest'
            example:
              plan: Premium
      responses:
        '200':
          description: Abonnement mis a jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
  /api/rest/succhero/chat:
    post:
      tags: [SUCCHERO]
      summary: Discuter avec SUCCHERO
      description: SUCCHERO repond uniquement aux questions qui concernent PrimeStream.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            example:
              message: Comment acceder a PrimeStream depuis un autre PC ?
      responses:
        '200':
          description: Reponse de SUCCHERO
          content:
            application/json:
              schema:
                type: object
                properties:
                  reply: { type: string }
...
  /api/rest/succhero/notify:
    post:
      tags: [SUCCHERO]
      summary: Activer une notification quand un film sera disponible
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvailabilityRequest'
            example:
              searchedTitle: The Wicked
              preferredGenre: Fantastique
      responses:
        '200':
          description: Demande creee ou deja existante
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id: { type: integer, format: int64 }
        username: { type: string }
        email: { type: string }
        fullName: { type: string }
        roles:
          type: array
          items: { type: string }
        primeMember: { type: boolean }
        primePlan: { type: string }
        primeSubscriptionEndsAt: { type: string, format: date-time }
        moodMode: { type: string }
        preferredGenre: { type: string }
...
    PreferenceRequest:
      type: object
      properties:
        moodMode: { type: string, enum: [classic, romance, action, family, science, afrique] }
        preferredGenre: { type: string }
        preferredLanguage: { type: string }
        preferredQuality: { type: string }
        maturityLevel: { type: string }
    SubscriptionRequest:
      type: object
      properties:
        plan: { type: string, enum: [Gratuit, Standard, Premium] }
...
    ChatRequest:
      type: object
      properties:
        message: { type: string }
    AvailabilityRequest:
      type: object
      properties:
        searchedTitle: { type: string }
        preferredGenre: { type: string }
...